refactor: move extra path to utils and changes paths objects
