+++ tools/serverpod_cli/lib/src/generator/open_api/helpers/extensions.dart
