+++ tools/serverpod_cli/lib/src/generator/open_api/objects/paths.dart
@@ -18,7 +18,7 @@ class PathsObject {
   });
 
   Map<String, dynamic> toJson() {
-    return {'/$pathName': path ?? {}};
+    return {pathName: path!.toJson()};
   }
 }
